refactor: convert annif.vocab submodule into a directory
