+++ annif/vocab/__init__.py
@@ -1,0 +1,12 @@
+"""Annif vocabulary functionality"""
+
+from .subject_index import SubjectIndexFile, SubjectIndexFilter
+from .types import SubjectIndex
+from .vocab import AnnifVocabulary
+
+__all__ = [
+    "AnnifVocabulary",
+    "SubjectIndex",
+    "SubjectIndexFile",
+    "SubjectIndexFilter",
+]
